Refactor fire page
