+++ Site/bc_weather_station_dashboard/website/static/WidgetJavaScript/fwi.js
